Create helpers and constants modules for s3 directory
This is not necessary right now, however when uploading files that have been created using the web scraper version 2.0 the process will be more complicated and separate modules will be beneficial
