+++ chinese-news-aggregator/etl/s3/constants.py
@@ -1,0 +1,3 @@
+from pathlib import Path
+
+DATA_DIRECTORY = Path.home() / "OneDrive/2021/Programming/python/chinese-news-aggregator/chinese-news-aggregator/data"
